Refactor document upload logic to StorageRepository
Moved application document upload methods from ApplicationRepository to a new StorageRepository. Deleted UploadApplicationDocument and UploadApplicationDocuments use cases, and updated ApplicationViewModel to remove their usage. This change improves separation of concerns and prepares for future enhancements to document handling.
